requirements.txt
discord.py==2.3.2
pymongo>=4.6
pillow>=10.0
requests>=2.31
python-dotenv>=1.0
